refactor home map page + add documentation
